Estructura unificada: Front, Back, y script de tests automatizados configurados
