--- index-AJM.js
@@ -1,33 +1,0 @@
-//Array con los 10 datos de ejemplo de mi ficha
-const datos = [
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "Australia", dead_weight_tons: 12.6, percentage_of_total_fleet: 9.796, number_of_ships: 1 },
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "China, Hong Kong SAR", dead_weight_tons: 79.56, percentage_of_total_fleet: 61.472, number_of_ships: 8 },
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "Colombia", dead_weight_tons: 3.6, percentage_of_total_fleet: 2.781, number_of_ships: 1 },
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "Cyprus", dead_weight_tons: 16.278, percentage_of_total_fleet: 12.576, number_of_ships: 2 },
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "Denmark", dead_weight_tons: 35.395, percentage_of_total_fleet: 27.345, number_of_ships: 4 },
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "Estonia", dead_weight_tons: 27.789, percentage_of_total_fleet: 21.469, number_of_ships: 7 },
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "Germany", dead_weight_tons: 12232.14, percentage_of_total_fleet: 9450.155, number_of_ships: 1056 },
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "Greece", dead_weight_tons: 7.496, percentage_of_total_fleet: 5.791, number_of_ships: 2 },
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "Iceland", dead_weight_tons: 58.93, percentage_of_total_fleet: 45.53, number_of_ships: 7 },
-    { year: 2014, flag_of_registration_label: "Antigua and Barbuda", beneficial_ownership_label: "Latvia", dead_weight_tons: 36.465, percentage_of_total_fleet: 28.172, number_of_ships: 10 }
-];
-
-//Valor del campo de información geográfica a buscar
-const paisBuscado = "Antigua and Barbuda";
-
-//Uso filter() para obtener el subconjunto de filas de ese país
-const datosFiltrados = datos.filter(fila => fila.flag_of_registration_label === paisBuscado);
-
-//Uso map() para extraer la última columna y reduce() para sumar todos sus valores
-const sumaBarcos = datosFiltrados
-    .map(fila => fila.number_of_ships)
-    .reduce((acumulador, valorActual) => acumulador + valorActual, 0);
-
-//Calculo la media
-let mediaBarcos = 0;
-if (datosFiltrados.length > 0) {
-    mediaBarcos = sumaBarcos / datosFiltrados.length;
-}
-
-//Resultado por consola
-console.log(`La media de 'number_of_ships' para '${paisBuscado}' es: ${mediaBarcos}`);
