+++ package.json
@@ -15,8 +15,12 @@
   "type": "module",
   "main": "index-AJM.js",
   "scripts": {
-    "start": "node index.js",
-    "test-AJM": "newman run tests/AJM-collection-POSTMAN.json --env-var \"url=http://localhost:3000\""
+    "start": "node src/back/index.js",
+    "test-AJM": "newman run tests/AJM-collection-POSTMAN.json --env-var \"url=http://localhost:3000\"",
+    "test-JDD": "echo \"Test de Jorge pendiente de subir\"",
+    "test-MZA": "echo \"Test de Zhillan pendiente de subir\"",
+    "test-all": "npm run test-AJM && npm run test-JDD && npm run test-MZA",
+    "test": "start-server-and-test start http://localhost:3000 test-all"
   },
   "dependencies": {
     "cool-ascii-faces": "^1.3.4",
@@ -24,6 +28,7 @@
     "nedb": "^1.8.0"
   },
   "devDependencies": {
-    "newman": "^6.2.2"
+    "newman": "^6.2.2",
+    "start-server-and-test": "^2.1.5"
   }
 }
