+++ public/about/index.html
@@ -4,29 +4,69 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>About - SOS2526-28</title>
+    <link rel="stylesheet" href="/static/style.css">
 </head>
 <body>
-    <h2>SOS2526-28</h2>
-    
-    <ul>
-        <li><strong>Team</strong>
-            <ul>
-                <li>Jorge Dayoub Dayoub (<a href="https://github.com/jorgedx2">https://github.com/jorgedx2</a>)</li>
-                <li>Alberto Jiménez Mateos (<a href="https://github.com/AlbertoJimenezMateos">https://github.com/AlbertoJimenezMateos</a>)</li>
-                <li>Muhammad Zhillan Averous (<a href="https://github.com/zaverous">https://github.com/zaverous</a>)</li>
-            </ul>
-        </li>
-        <li><strong>Project description:</strong> Our sources of information are a jumble of unrelated data; our intention will be to find a connection between the three data sources.</li>
-        <li><strong>Repository:</strong> <a href="https://github.com/gti-sos/SOS2526-28">gti-sos/SOS2526-28</a></li>
+    <div class="container">
+        <h1>ℹ️ About - SOS2526-28</h1>
         
-        <li><strong>URL de Despliegue:</strong> <a href="https://sos2526-28.onrender.com">https://sos2526-28.onrender.com</a></li>
-        <li><strong>APIs:</strong>
-            <ul>
-                <li>API Alberto: <a href="https://sos2526-28.onrender.com/api/v1/beneficial-ownership-merchant-fleets"> https://sos2526-28.onrender.com/api/v1/beneficial-ownership-merchant-fleets</a></li>
-                <li>API Jorge: <a href="https://sos2526-28.onrender.com/api/v1/deliberate-violence-against-civilians-events-worldwide">https://sos2526-28.onrender.com/api/v1/deliberate-violence-against-civilians-events-worldwide</a></li>
-                <li>API Zhillan: <a href="https://sos2526-28.onrender.com/api/v1/company-esg-scores-financial-performances">https://sos2526-28.onrender.com/api/v1/company-esg-scores-financial-performances</a></li>
+        <div class="endpoint-group">
+            <h3>👥 Team</h3>
+            <ul class="api-list">
+                <li class="api-item">
+                    <span class="api-title">Alberto Jiménez Mateos</span>
+                    <a href="https://github.com/AlbertoJimenezMateos" target="_blank">https://github.com/AlbertoJimenezMateos</a>
+                </li>
+                <li class="api-item">
+                    <span class="api-title">Jorge Dayoub Dayoub</span>
+                    <a href="https://github.com/jorgedx2" target="_blank">https://github.com/jorgedx2</a>
+                </li>
+                <li class="api-item">
+                    <span class="api-title">Muhammad Zhillan Averous</span>
+                    <a href="https://github.com/zaverous" target="_blank">https://github.com/zaverous</a>
+                </li>
             </ul>
-        </li>
-    </ul>
+        </div>
+
+        <div class="endpoint-group">
+            <h3>📝 Project description</h3>
+            <div class="api-item">
+                <p>Our sources of information are a jumble of unrelated data; our intention will be to find a connection between the three data sources.</p>
+            </div>
+        </div>
+
+        <div class="endpoint-group">
+            <h3>🔗 Links & APIs</h3>
+            <ul class="api-list">
+                <li class="api-item">
+                    <span class="api-title">Repository</span>
+                    <a href="https://github.com/gti-sos/SOS2526-28" target="_blank">gti-sos/SOS2526-28</a>
+                </li>
+                <li class="api-item">
+                    <span class="api-title">Deployment URL</span>
+                    <a href="https://sos2526-28.onrender.com" target="_blank">https://sos2526-28.onrender.com</a>
+                </li>
+                <li class="api-item">
+                    <span class="api-title">API Alberto</span>
+                    <a href="https://sos2526-28.onrender.com/api/v1/beneficial-ownership-merchant-fleets" target="_blank">https://sos2526-28.onrender.com/api/v1/beneficial-ownership-merchant-fleets</a>
+                </li>
+                <li class="api-item">
+                    <span class="api-title">API Jorge</span>
+                    <a href="https://sos2526-28.onrender.com/api/v1/deliberate-violence-against-civilians-events-worldwide" target="_blank">https://sos2526-28.onrender.com/api/v1/deliberate-violence-against-civilians-events-worldwide</a>
+                </li>
+                <li class="api-item">
+                    <span class="api-title">API Zhillan</span>
+                    <a href="https://sos2526-28.onrender.com/api/v1/company-esg-scores-financial-performances" target="_blank">https://sos2526-28.onrender.com/api/v1/company-esg-scores-financial-performances</a>
+                </li>
+            </ul>
+        </div>
+
+        <div class="endpoint-group">
+            <div class="info-links">
+                <a href="/" class="info-link">⬅️ Volver al Inicio</a>
+            </div>
+        </div>
+
+    </div>
 </body>
 </html>
