--- public/index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="es">
-<head>
-    <meta charset="UTF-8">
-    <title>SOS2526-28 - Panel de Control</title>
-    <style>
-        body { font-family: sans-serif; line-height: 1.6; padding: 20px; max-width: 800px; margin: auto; }
-        h1 { color: #2c3e50; border-bottom: 2px solid #eee; }
-        .endpoint-group { background: #f9f9f9; padding: 15px; border-radius: 8px; margin-bottom: 10px; }
-        a { color: #3498db; text-decoration: none; }
-        a:hover { text-decoration: underline; }
-    </style>
-</head>
-<body>
-    <h1>🚀 SOS2526-28 en funcionamiento</h1>
-    
-   <div class="endpoint-group">
-    <h3>📊 APIs Disponibles:</h3>
-    <ul>
-        <li><b>Beneficial Ownership Merchant Fleets:</b> <a href="/api/v1/beneficial-ownership-merchant-fleets">/api/v1/beneficial-ownership-merchant-fleets</a></li>
-        <li><b>Deliberate Violence Against Civilians Events Worldwide:</b> <a href="/api/v1/deliberate-violence-against-civilians-events-worldwide">/api/v1/deliberate-violence-against-civilians-events-worldwide</a></li>
-        <li><b>Company ESG Scores Financial Performances:</b> <a href="/api/v1/company-esg-scores-financial-performances">/api/v1/company-esg-scores-financial-performances</a></li>
-    </ul>
-    </div>
-
-    <div class="endpoint-group">
-        <h3>ℹ️ Información:</h3>
-        <p><a href="/about">Sobre nosotros (About)</a></p>
-        <p><a href="/api/v1/docs">Documentación de la API (Postman)</a></p>
-    </div>
-</body>
-</html>
