+++ src/front/index.html
@@ -1,0 +1,51 @@
+<!DOCTYPE html>
+<html lang="es">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>SOS2526-28 - Panel de Control</title>
+    <link rel="stylesheet" href="static/style.css">
+</head>
+<body>
+    <div class="container">
+        <h1>🚀 SOS2526-28</h1>
+        
+        <div class="endpoint-group">
+            <h3>📊 APIs</h3>
+            <ul class="api-list">
+                <li class="api-item">
+                    <span class="api-title"><h3>BENEFICIAL OWNERSHIP MERCHANT FLEETS</h3></span>
+                    <span class="api-title">API (RENDER)</span>
+                    <a href="/api/v1/beneficial-ownership-merchant-fleets">https://sos2526-28.onrender.com/api/v1/beneficial-ownership-merchant-fleets</a>
+                    <span class="api-title">DOCUMENTACIÓN POSTMAN</span>
+                    <a href="https://documenter.getpostman.com/view/52406665/2sBXigLsxq">https://documenter.getpostman.com/view/52406665/2sBXigLsxq</a><br>
+                    <span class="api-title">(developed by Alberto Jiménez Mateos)</span>
+                </li>
+                <li class="api-item">
+                    <span class="api-title"><h3>DELIBERATE VIOLENCE AGAINST CIVILIANS EVENTS WORLDWIDE</h3></span>
+                    <span class="api-title">API (RENDER)</span>
+                    <a href="/api/v1/deliberate-violence-against-civilians-events-worldwide">https://sos2526-28.onrender.com/api/v1/deliberate-violence-against-civilians-events-worldwide</a>
+                    <span class="api-title">DOCUMENTACIÓN POSTMAN</span>
+                    <a href="https://sos2526-28.onrender.com/api/v1/deliberate-violence-against-civilians-events-worldwide">https://sos2526-28.onrender.com/api/v1/deliberate-violence-against-civilians-events-worldwide</a><br>
+                    <span class="api-title">(developed by Jorge Dayoub Dayoub)</span>
+                </li>
+                <li class="api-item">
+                    <span class="api-title"><h3>COMPANY ESG SCORES FINANCIAL PERFORMANCES</h3></span>
+                    <span class="api-title">API (RENDER)</span>
+                    <a href="/api/v1/company-esg-scores-financial-performances">https://sos2526-28.onrender.com/api/v1/company-esg-scores-financial-performances</a>
+                    <span class="api-title">DOCUMENTACIÓN POSTMAN</span>
+                    <a href="/api/v1/company-esg-scores-financial-performances">https://sos2526-28.onrender.com/api/v1/company-esg-scores-financial-performances</a><br>
+                    <span class="api-title">(developed by Muhammad Zhillan Averous)</span>
+                </li>
+            </ul>
+        </div>
+
+        <div class="endpoint-group">
+            <h3>ℹ️ INFORMACIÓN</h3>
+            <div class="info-links">
+                <a href="/about" class="info-link">Sobre nosotros (About)</a>
+            </div>
+        </div>
+    </div>
+</body>
+</html>
